Refactor validators
RISDEV-0000
