+++ backend/src/main/java/de/bund/digitalservice/ris/norms/application/service/ModificationValidator.java
@@ -1,15 +1,10 @@
 package de.bund.digitalservice.ris.norms.application.service;
 
-import de.bund.digitalservice.ris.norms.application.port.output.LoadNormPort;
-import de.bund.digitalservice.ris.norms.domain.entity.*;
-import de.bund.digitalservice.ris.norms.utils.XmlMapper;
-import de.bund.digitalservice.ris.norms.utils.exceptions.XmlContentException;
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
+import de.bund.digitalservice.ris.norms.domain.entity.CharacterRange;
+import de.bund.digitalservice.ris.norms.domain.entity.Mod;
+import de.bund.digitalservice.ris.norms.domain.entity.Norm;
+import de.bund.digitalservice.ris.norms.domain.entity.TextualMod;
+import de.bund.digitalservice.ris.norms.utils.exceptions.ValidationException;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.w3c.dom.Node;
@@ -17,373 +12,66 @@ import org.w3c.dom.Node;
 /** */
 @Service
 public class ModificationValidator {
-  private final LoadNormPort loadNormPort;
-  private final LoadZf0Service loadZf0Service;
-
-  public ModificationValidator(LoadNormPort loadNormPort, LoadZf0Service loadZf0Service) {
-    this.loadNormPort = loadNormPort;
-    this.loadZf0Service = loadZf0Service;
-  }
-
-  /**
-   * Checks an amending law xml for consistency errors.
-   *
-   * @param amendingNorm the amending norm to be checked
-   */
-  public void validate(Norm amendingNorm) {
-    destinationIsSet(amendingNorm);
-    destinationEliIsConsistent(amendingNorm);
-    destinationHrefIsConsistent(amendingNorm);
-    checkAllMods(amendingNorm);
-  }
-
-  private void checkAllMods(Norm amendingNorm) {
-    String amendingNormEli = amendingNorm.getEli();
-    List<Article> articles =
-        amendingNorm.getArticles().stream()
-            .filter(
-                article -> {
-                  String articleRefersTo = article.getRefersToOrThrow();
-                  return Objects.equals(articleRefersTo, "hauptaenderung");
-                })
-            .toList();
-
-    List<Mod> mods = articles.stream().map(Article::getMods).flatMap(List::stream).toList();
-
-    mods.forEach(mod -> validateSubstitutionMod(amendingNormEli, mod));
-  }
-
-  /**
-   * Throws an error if any of the articles of the passed amendingNorm has an empty affected
-   * document Eli. The error message contains a comma separated list of all article eIds, that are
-   * affected.
-   *
-   * @param amendingNorm the amending law to be checked
-   */
-  public void destinationIsSet(Norm amendingNorm) {
-    validateActiveModificationsEli(amendingNorm);
-    validateArticlesEli(amendingNorm);
-  }
-
-  private void validateActiveModificationsEli(Norm amendingNorm) {
-    amendingNorm
-        .getMeta()
-        .getAnalysis()
-        .map(Analysis::getActiveModifications)
-        .orElse(Collections.emptyList())
-        .forEach(
-            tm ->
-                tm.getDestinationHref()
-                    .orElseThrow(
-                        () -> {
-                          String amendingNormEli = amendingNorm.getEli();
-                          return new XmlContentException(
-                              "For norm with Eli (%s): ActiveModification Destination Href is empty where textualMod eId is %s"
-                                  .formatted(
-                                      amendingNormEli, getTextualModEId(amendingNormEli, tm)),
-                              null);
-                        })
-                    .getEli()
-                    .orElseThrow(
-                        () -> {
-                          String amendingNormEli = amendingNorm.getEli();
-                          return new XmlContentException(
-                              "For norm with Eli (%s): ActiveModification Destination Href holds an empty (more general: invalid) Eli where textualMod eId is %s"
-                                  .formatted(
-                                      amendingNormEli, getTextualModEId(amendingNormEli, tm)),
-                              null);
-                        }));
-  }
-
-  private void validateArticlesEli(Norm amendingNorm) {
-    amendingNorm.getArticles().stream()
-        .filter(
-            article -> {
-              String articleRefersTo = article.getRefersToOrThrow();
-              return Objects.equals(articleRefersTo, "hauptaenderung");
-            })
-        .forEach(
-            article -> {
-              String amendingNormEli = amendingNorm.getEli();
-              String articleEId = article.getEidOrThrow();
-              validateAknModEli(amendingNormEli, article, articleEId);
-              validateAffectedDocumentEli(amendingNormEli, article, articleEId);
-            });
-  }
-
-  private void validateAknModEli(String amendingNormEli, Article article, String articleEId) {
-    List<Mod> mods = article.getModsOrThrow();
-    mods.forEach(
-        mod -> {
-          Optional<String> eli = getModTargetHref(amendingNormEli, mod, articleEId).getEli();
-          if (eli.isEmpty()) {
-            throw new XmlContentException(
-                "For norm with Eli (%s): The Eli in aknMod href is empty in article with eId %s"
-                    .formatted(amendingNormEli, articleEId),
-                null);
-          }
-        });
-  }
-
-  private void validateAffectedDocumentEli(
-      String amendingNormEli, Article article, String articleEId) {
-    if (article.getAffectedDocumentEli().isEmpty()) {
-      throw new XmlContentException(
-          "For norm with Eli (%s): AffectedDocument href is empty in article with eId %s"
-              .formatted(amendingNormEli, articleEId),
-          null);
-    }
-  }
-
-  /**
-   * Checks whether a reference to a target law in an amending law is consistent in
-   * akn:activeModifications - destination and akn:mod
-   *
-   * @param amendingNorm the amending law to be checked
-   */
-  public void destinationEliIsConsistent(Norm amendingNorm) {
-    Set<String> affectedDocumentElis =
-        amendingNorm.getArticles().stream()
-            .map(Article::getAffectedDocumentEli)
-            .flatMap(Optional::stream)
-            .collect(Collectors.toSet());
-
-    Set<String> activeModificationsDestinationElis =
-        amendingNorm
-            .getMeta()
-            .getAnalysis()
-            .map(Analysis::getActiveModifications)
-            .orElse(Collections.emptyList())
-            .stream()
-            .map(TextualMod::getDestinationHref)
-            .flatMap(Optional::stream)
-            .map(Href::getEli)
-            .flatMap(Optional::stream)
-            .collect(Collectors.toSet());
-
-    Set<String> aknModElis =
-        amendingNorm.getArticles().stream()
-            .map(Article::getMods)
-            .flatMap(List::stream)
-            .map(Mod::getTargetHref)
-            .flatMap(Optional::stream)
-            .map(Href::getEli)
-            .flatMap(Optional::stream)
-            .collect(Collectors.toSet());
-
-    if (!affectedDocumentElis.equals(activeModificationsDestinationElis)
-        && !activeModificationsDestinationElis.equals(aknModElis))
-      throw new XmlContentException(
-          "For norm with Eli (%s): Elis are not consistent".formatted(amendingNorm.getEli()), null);
-  }
-
-  /**
-   * Checks whether a reference to a target law in an amending law is consistent in
-   * akn:activeModifications - destination and akn:mod
-   *
-   * @param amendingNorm the amending law to be checked
-   */
-  public void destinationHrefIsConsistent(Norm amendingNorm) {
-    Set<Href> activeModificationsDestinationHrefs =
-        amendingNorm
-            .getMeta()
-            .getAnalysis()
-            .map(Analysis::getActiveModifications)
-            .orElse(Collections.emptyList())
-            .stream()
-            .map(TextualMod::getDestinationHref)
-            .flatMap(Optional::stream)
-            .collect(Collectors.toSet());
-
-    Set<Href> aknModHrefs =
-        amendingNorm.getArticles().stream()
-            .map(Article::getMods)
-            .flatMap(List::stream)
-            .map(Mod::getTargetHref)
-            .flatMap(Optional::stream)
-            .collect(Collectors.toSet());
-    if (!activeModificationsDestinationHrefs.equals(aknModHrefs))
-      throw new XmlContentException(
-          "For norm with Eli (%s): Eids are not consistent".formatted(amendingNorm.getEli()), null);
-  }
 
   /**
    * Checks if a substitution mod is consistent
    *
-   * @param amendingLawEli the amending Norm Eli
-   * @param mod the amending law mod to be checked
+   * @param zf0Norm the zf0Norm
+   * @param activeMod the active mod of the amending law
    */
-  public void validateSubstitutionMod(String amendingLawEli, Mod mod) {
-    if (mod.usesQuotedText()) {
-      validateQuotedTextSubstitutions(amendingLawEli, mod);
+  public void validateSubstitutionMod(final Norm zf0Norm, final Mod activeMod) {
+
+    final String modEId = activeMod.getMandatoryEid();
+
+    final TextualMod affectedPassiveMod =
+        zf0Norm.getMeta().getOrCreateAnalysis().getPassiveModifications().stream()
+            .filter(f -> f.getSourceHref().orElseThrow().getEId().orElseThrow().equals(modEId))
+            .findFirst()
+            .orElseThrow();
+
+    final String targetNodeEid =
+        affectedPassiveMod.getDestinationHref().orElseThrow().getEId().orElseThrow();
+    final Node zf0TargetedNode =
+        zf0Norm
+            .getNodeByEId(targetNodeEid)
+            .orElseThrow(
+                () ->
+                    new ValidationException(
+                        "Target node with eid %s not present".formatted(targetNodeEid)));
+
+    if (activeMod.usesQuotedText()) {
+      validateQuotedText(
+          affectedPassiveMod,
+          StringUtils.normalizeSpace(activeMod.getMandatoryOldText()),
+          StringUtils.normalizeSpace(zf0TargetedNode.getTextContent()));
     }
     // other case <akn:quotedStructure>
   }
 
-  /**
-   * Checks if the text to be replaced is present in the target law
-   *
-   * @param amendingNormEli the amending Norm Eli
-   * @param mod the amending law mod to be checked
-   */
-  private void validateQuotedTextSubstitutions(String amendingNormEli, Mod mod) {
-    String modEId =
-        mod.getEid()
-            .orElseThrow(
-                () ->
-                    new XmlContentException(
-                        "Eid in mod %s is empty".formatted(XmlMapper.toString(mod.getNode())),
-                        null));
-    Href articleModTargetHref = getModTargetHref(amendingNormEli, mod, modEId);
-    Norm amendingLaw =
-        loadNormPort
-            .loadNorm(new LoadNormPort.Command(amendingNormEli))
-            .orElseThrow(() -> new XmlContentException("Could not load norm", null));
-    Norm zf0Norm = loadZf0Service.loadZf0(amendingLaw, mod);
-
-    String articleModTargetHrefEId =
-        articleModTargetHref
-            .getEId()
-            .orElseThrow(
-                () ->
-                    new XmlContentException(
-                        "For norm with Eli (%s): The eId in mod href is empty in article with eId %s"
-                            .formatted(amendingNormEli, modEId),
-                        null));
-
-    Node zf0TargetNode =
-        getTargetNodeFromZF0Norm(
-            amendingNormEli, zf0Norm, articleModTargetHrefEId, amendingNormEli, modEId);
-
-    // normalizeSpace removes double spaces and new lines
-    String targetParagraphOldText = StringUtils.normalizeSpace(zf0TargetNode.getTextContent());
-
-    String amendingNormOldText =
-        mod.getOldText()
-            .orElseThrow(
-                () ->
-                    new XmlContentException(
-                        "For norm with Eli (%s): quotedText[1] (the old, to be replaced, text) is empty in article mod with eId %s"
-                            .formatted(amendingNormEli, modEId),
-                        null));
-
-    // normalizeSpace removes double spaces and new lines
-    amendingNormOldText = StringUtils.normalizeSpace(amendingNormOldText);
-
-    validateQuotedText(
-        amendingNormEli,
-        amendingNormOldText,
-        targetParagraphOldText,
-        modEId,
-        articleModTargetHref,
-        "The character range in mod href is empty in article with eId %s".formatted(modEId));
-  }
-
   private void validateQuotedText(
-      String eli,
-      String amendingNormOldText,
-      String targetParagraphOldText,
-      String modEId,
-      Href href,
-      String message) {
+      final TextualMod passiveMode, String amendingNormOldText, String targetParagraphOldText) {
 
-    CharacterRange characterRange =
-        href.getCharacterRange()
-            .orElseThrow(
-                () ->
-                    new XmlContentException(
-                        "For norm with Eli (%s): %s".formatted(eli, message), null));
+    final CharacterRange characterRange =
+        passiveMode
+            .getDestinationHref()
+            .orElseThrow()
+            .getCharacterRange()
+            .orElseThrow(() -> new ValidationException("Character range not present"));
 
-    if (!characterRange.isValidCharacterRange()) {
-      throw new XmlContentException(
-          "The range (%s) given at mod with eId %s is not valid".formatted(characterRange, modEId),
-          null);
-    }
+    if (!characterRange.isValidCharacterRange())
+      throw new ValidationException(
+          "The character range (%s) is not valid".formatted(characterRange));
 
-    int modStart = characterRange.getStart();
-    int modEnd = characterRange.getEnd();
+    final int modStart = characterRange.getStart();
+    final int modEnd = characterRange.getEnd();
 
-    validateStartIsBeforeEnd(eli, characterRange, modStart, modEnd, modEId);
-    checkIfReplacementEndIsWithinText(eli, targetParagraphOldText, modEnd, modEId);
-
-    String zf0NormOldText = targetParagraphOldText.substring(modStart, modEnd);
-    validateQuotedTextEquals(eli, zf0NormOldText, amendingNormOldText, modEId);
-  }
-
-  private void validateStartIsBeforeEnd(
-      String amendingNormEli, CharacterRange cr, int start, int end, String modEId) {
-    if (!cr.isEndGreaterStart())
-      throw new XmlContentException(
-          "For norm with Eli (%s): The character range in mod href is not valid in mod with eId %s. Make sure start is smaller than end %s < %s."
-              .formatted(amendingNormEli, modEId, start, end),
-          null);
-  }
-
-  private void checkIfReplacementEndIsWithinText(
-      String amendingNormEli, String targetParagraphText, int modEnd, String modEId) {
-    // counting is null based e.g. 0 is the position of the first character; spaces are counted
-    // see
+    // counting is null based e.g. 0 is the position of the first character; spaces are counted see
     // https://gitlab.opencode.de/bmi/e-gesetzgebung/ldml_de/-/blob/1.6/Spezifikation/LegalDocML.de_1.6.pdf?ref_type=tags page 85
-    if (targetParagraphText.length() < modEnd) {
-      throw new XmlContentException(
-          "For norm with Eli (%s): The character range in mod href is not valid (target paragraph is to short) in mod with eId %s"
-              .formatted(amendingNormEli, modEId),
-          null);
-    }
-  }
+    if (targetParagraphOldText.length() < modEnd)
+      throw new ValidationException("The character range not within length of target node content");
 
-  private void validateQuotedTextEquals(
-      String amendingNormEli, String zf0NormOldText, String amendingNormOldText, String modEId) {
-    if (!zf0NormOldText.equals(amendingNormOldText))
-      throw new XmlContentException(
-          "For norm with Eli (%s): The replacement text '%s' in the target law does not equal the replacement text '%s' in the mod with eId %s"
-              .formatted(amendingNormEli, zf0NormOldText, amendingNormOldText, modEId),
-          null);
-  }
-
-  private Node getTargetNodeFromZF0Norm(
-      String amendingNormEli,
-      Norm zf0Norm,
-      String targetHrefEId,
-      String affectedDocumentEli,
-      String modEId) {
-
-    validateNumberOfNodesWithEid(zf0Norm.getEli(), zf0Norm, targetHrefEId);
-
-    return zf0Norm
-        .getNodeByEId(targetHrefEId)
-        .orElseThrow(
-            () ->
-                new XmlContentException(
-                    "For norm with Eli (%s): Couldn't load target eId (%s) element in zf0 (%s) for mod with eId %s"
-                        .formatted(amendingNormEli, targetHrefEId, affectedDocumentEli, modEId),
-                    null));
-  }
-
-  private void validateNumberOfNodesWithEid(String eli, Norm norm, String eId) {
-    if (norm.getNumberOfNodesWithEid(eId) > 1) {
-      throw new XmlContentException(
-          "For norm with Eli (%s): Too many elements with the same eId %s.".formatted(eli, eId),
-          null);
-    }
-  }
-
-  private Href getModTargetHref(String eli, Mod m, String modEId) {
-    return m.getTargetHref()
-        .orElseThrow(
-            () ->
-                new XmlContentException(
-                    "For norm with Eli (%s): mod href is empty in article mod with eId %s"
-                        .formatted(eli, modEId),
-                    null));
-  }
-
-  private String getTextualModEId(String eli, TextualMod textualMod) {
-    return textualMod
-        .getEid()
-        .orElseThrow(
-            () ->
-                new XmlContentException(
-                    "For norm with Eli (%s): TextualMod eId empty.".formatted(eli), null));
+    if (!targetParagraphOldText.substring(modStart, modEnd).equals(amendingNormOldText))
+      throw new ValidationException(
+          "The character range does not resolve to the targeted text to be replaced");
   }
 }
